feat: Fix blog template styling. Add autolink headers.
From <https://www.gatsbyjs.com/plugins/gatsby-remark-autolink-headers/>
