+++ gatsby-config.js
@@ -78,6 +78,11 @@ module.exports = {
         path: `${__dirname}/src/markdown-pages`,
       },
     },
-    'gatsby-transformer-remark',
+    {
+      resolve: 'gatsby-transformer-remark',
+      options: {
+        plugins: ['gatsby-remark-autolink-headers'],
+      },
+    }
   ],
 }
