+++ package.json
@@ -77,6 +77,7 @@
     "gatsby-plugin-sharp": "^2.6.42",
     "gatsby-plugin-svgr": "^2.0.2",
     "gatsby-plugin-webpack-bundle-analyser-v2": "^1.1.13",
+    "gatsby-remark-autolink-headers": "^4.9.0",
     "gatsby-source-filesystem": "^2.8.1",
     "gatsby-transformer-remark": "^2.13.1",
     "gatsby-transformer-sharp": "^2.5.17",
@@ -128,6 +129,6 @@
     "typescript": "^4.0.3"
   },
   "engines": {
-    "node": ">= 10.14.2"
+    "node": ">= 14.15.0"
   }
 }
